refactor: expand plugin contract metadata
